+++ packages/svelte/src/internal/client/custom-element.js
@@ -1,5 +1,5 @@
 import { createClassComponent } from '../../legacy/legacy-client.js';
-import { render_effect, destroy_signal } from './runtime.js';
+import { render_effect, destroy_signal } from '../common/runtime.js';
 import { open, close } from './render.js';
 import { define_property } from './utils.js';
