+++ packages/svelte/src/internal/server/index.js
@@ -1,5 +1,5 @@
-import * as $ from '../client/runtime.js';
-import { set_is_ssr } from '../client/runtime.js';
+import * as $ from '../common/runtime.js';
+import { set_is_ssr } from '../common/runtime.js';
 import { is_promise } from '../common/index.js';
 import { subscribe_to_store } from '../../store/utils.js';
